requirements.txt
openai
requests
aiohttp
python-dotenv
rich
textstat
click
google-generativeai